+++ src/main/java/View/View.java
@@ -1,0 +1,9 @@
+package View;
+
+public class View {
+    // TODO Transfer all the views from Controller class here
+
+    public static void LoginView() {
+        // TODO: Implement this method
+    }
+}